GH-4492 route all deprecated endpoints through new controllers
